genAi-week02/requirements.txt
langchain
langchain-core
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv
